Spike for jooby plus X
This brings quite a few changes on the table:

1. Domain model as shared code
To avoid some duplication, the ebean based domain model will be in the new `shared` project.
This comes with quite some changes on the existing Spring app and in the build scripts.

2. Jooby spike works!
All code needed to have the same functionallity as we have with Spring is in here now.

3. Koin as DI framework
We will use Koin for all framework approaches but Spring. This commit show how it is done for Jooby already 🎉