Docker-fy Jooby
A bit rude, but it works, the Docker-isation of the Jooby demo.

Use the _Dockerfile_ in _jooby/_ to create an image. The image itself accepts `DB_NAME`, `DB_HOST`, `DB_USERNAME` and `DB_PASSWORD` as variable and generates the relevant ebean configuration file.
The current build of the image is also available on [Docker Hub](https://hub.docker.com/r/daincredibleholg/fosdem21-demo-jooby).