Fixes
First of all, this fixes a minor design bug in the database design. An image filename is not having a unique index anymore but has a normal index on it now.

Also, a small change to the DataDog configuration comes in with this commit. It should hopefully enable some more JVM insight gathering.

Last but not least, the _ebean.mf_ is back. Looks like ebean needs it after all.